LPS-63470 move to messaging package
