+++ modules/apps/platform/portal-configuration/portal-configuration-cluster/src/main/java/com/liferay/portal/configuration/cluster/messaging/DestinationNames.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.configuration.cluster;
+package com.liferay.portal.configuration.cluster.messaging;
 
 /**
  * @author Michael C. Han
